+++ Rant/Properties/AssemblyInfo.cs
@@ -67,4 +67,4 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
 
-[assembly: AssemblyVersion("3.0.3")]
+[assembly: AssemblyVersion("3.0.4")]